// auth/features/ory-elements/types.ts
export type OryNodeCheckboxInputProps = {
  name: string
  /* eslint-disable-next-line @typescript-eslint/no-explicit-any */
  onChange: (event: any) => void
  disabled?: boolean
  type: "checkbox"
  checked: boolean
  value: string | number | readonly string[] | undefined
  /* eslint-disable-next-line @typescript-eslint/no-explicit-any */
  ref: (instance: any) => void
}

export type OryNodeCheckboxProps = {
  /** @deprecated - use node.attributes */
  attributes: UiNodeInputAttributes
  node: UiNodeInput
  /** @deprecated - use inputProps.onClick */
  onClick: MouseEventHandler
  inputProps: OryNodeCheckboxInputProps
}